requirements.txt
arxiv==2.2.0
Flask==3.1.2
langchain==0.3.27
langchain_community==0.3.30
langchain_core==0.3.78
langchain_google_genai==2.1.12
langchain_text_splitters==0.3.11
pdfplumber==0.11.7
python-dotenv==1.1.1
torch==2.8.0
transformers==4.57.0
